+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   New event `PLAYER_DEAD`
+
+### Changed
+
+-   Code refactor
+-   Removed `init.lua` and `settings.lua`
+
 ## [1.0.0] - 2024-01-02
 
 ### Added